config: delete helper fields from TestConfiguration
The processing required to fill the helper fields
PodsUnderTestLabeslObject and OperatorsUnderTestLabelsObject is moved to
the autodiscover package. The fields are deleted from the struct that
holds the config (TestConfiguration).
