+++ pkg/autodiscover/autodiscover.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"regexp"
 	"time"
 
 	configv1 "github.com/openshift/api/config/v1"
@@ -94,6 +95,11 @@ type DiscoveredTestData struct {
 	CollectorAppPassword   string
 }
 
+type labelObject struct {
+	LabelKey   string
+	LabelValue string
+}
+
 var data = DiscoveredTestData{}
 
 func buildLabelName(labelPrefix, labelName string) string {
@@ -122,6 +128,26 @@ func warnDeprecation(config *configuration.TestConfiguration) {
 	}
 }
 
+const labelRegex = `(\S*)\s*:\s*(\S*)`
+const labelRegexMatches = 3
+
+func createLabels(labelStrings []string) (labelObjects []labelObject) {
+	for _, label := range labelStrings {
+		r := regexp.MustCompile(labelRegex)
+
+		values := r.FindStringSubmatch(label)
+		if len(values) != labelRegexMatches {
+			logrus.Errorf("failed to parse label=%s, will not be used!, ", label)
+			continue
+		}
+		var aLabel labelObject
+		aLabel.LabelKey = values[1]
+		aLabel.LabelValue = values[2]
+		labelObjects = append(labelObjects, aLabel)
+	}
+	return labelObjects
+}
+
 // DoAutoDiscover finds objects under test
 //
 //nolint:funlen
@@ -134,18 +160,21 @@ func DoAutoDiscover(config *configuration.TestConfiguration) DiscoveredTestData 
 		logrus.Fatalf("Failed to retrieve storageClasses - err: %v", err)
 	}
 
+	podsUnderTestLabelsObjects := createLabels(config.PodsUnderTestLabels)
+	operatorsUnderTestLabelsObjects := createLabels(config.OperatorsUnderTestLabels)
+
 	// prints warning about deprecated labels
 	warnDeprecation(config)
 	// consolidate pods labels
 	for _, aLabel := range config.TargetPodLabels {
 		key, value := buildLabelKeyValue(aLabel)
-		config.PodsUnderTestLabelsObjects = append(config.PodsUnderTestLabelsObjects, configuration.LabelObject{LabelKey: key, LabelValue: value})
+		podsUnderTestLabelsObjects = append(podsUnderTestLabelsObjects, labelObject{LabelKey: key, LabelValue: value})
 	}
 	// adds DEPRECATED hardcoded operator label
-	config.OperatorsUnderTestLabelsObjects = append(config.OperatorsUnderTestLabelsObjects, configuration.LabelObject{LabelKey: deprecatedHardcodedOperatorLabelName, LabelValue: deprecatedHardcodedOperatorLabelValue})
+	operatorsUnderTestLabelsObjects = append(operatorsUnderTestLabelsObjects, labelObject{LabelKey: deprecatedHardcodedOperatorLabelName, LabelValue: deprecatedHardcodedOperatorLabelValue})
 
-	logrus.Infof("parsed pods under test labels: %+v", config.PodsUnderTestLabelsObjects)
-	logrus.Infof("parsed operators under test labels: %+v", config.OperatorsUnderTestLabelsObjects)
+	logrus.Infof("parsed pods under test labels: %+v", podsUnderTestLabelsObjects)
+	logrus.Infof("parsed operators under test labels: %+v", operatorsUnderTestLabelsObjects)
 
 	data.AllNamespaces, _ = getAllNamespaces(oc.K8sClient.CoreV1())
 	data.AllSubscriptions = findSubscriptions(oc.OlmClient, []string{""})
@@ -153,9 +182,9 @@ func DoAutoDiscover(config *configuration.TestConfiguration) DiscoveredTestData 
 	data.AllInstallPlans = getAllInstallPlans(oc.OlmClient)
 	data.AllCatalogSources = getAllCatalogSources(oc.OlmClient)
 	data.Namespaces = namespacesListToStringList(config.TargetNameSpaces)
-	data.Pods, data.AllPods = findPodsByLabel(oc.K8sClient.CoreV1(), config.PodsUnderTestLabelsObjects, data.Namespaces)
+	data.Pods, data.AllPods = findPodsByLabel(oc.K8sClient.CoreV1(), podsUnderTestLabelsObjects, data.Namespaces)
 	data.AbnormalEvents = findAbnormalEvents(oc.K8sClient.CoreV1(), data.Namespaces)
-	debugLabels := []configuration.LabelObject{{LabelKey: debugHelperPodsLabelName, LabelValue: debugHelperPodsLabelValue}}
+	debugLabels := []labelObject{{LabelKey: debugHelperPodsLabelName, LabelValue: debugHelperPodsLabelValue}}
 	debugNS := []string{config.DebugDaemonSetNamespace}
 	data.DebugPods, _ = findPodsByLabel(oc.K8sClient.CoreV1(), debugLabels, debugNS)
 	data.ResourceQuotaItems, err = getResourceQuotas(oc.K8sClient.CoreV1())
@@ -172,7 +201,7 @@ func DoAutoDiscover(config *configuration.TestConfiguration) DiscoveredTestData 
 	}
 	data.Crds = FindTestCrdNames(config.CrdFilters)
 	data.ScaleCrUnderTest = GetScaleCrUnderTest(data.Namespaces, data.Crds)
-	data.Csvs = findOperatorsByLabel(oc.OlmClient, config.OperatorsUnderTestLabelsObjects, config.TargetNameSpaces)
+	data.Csvs = findOperatorsByLabel(oc.OlmClient, operatorsUnderTestLabelsObjects, config.TargetNameSpaces)
 	data.Subscriptions = findSubscriptions(oc.OlmClient, data.Namespaces)
 	data.HelmChartReleases = getHelmList(oc.RestConfig, data.Namespaces)
 
@@ -194,8 +223,8 @@ func DoAutoDiscover(config *configuration.TestConfiguration) DiscoveredTestData 
 	data.OCPStatus = compatibility.DetermineOCPStatus(openshiftVersion, time.Now())
 
 	data.K8sVersion = k8sVersion.GitVersion
-	data.Deployments = findDeploymentByLabel(oc.K8sClient.AppsV1(), config.PodsUnderTestLabelsObjects, data.Namespaces)
-	data.StatefulSet = findStatefulSetByLabel(oc.K8sClient.AppsV1(), config.PodsUnderTestLabelsObjects, data.Namespaces)
+	data.Deployments = findDeploymentByLabel(oc.K8sClient.AppsV1(), podsUnderTestLabelsObjects, data.Namespaces)
+	data.StatefulSet = findStatefulSetByLabel(oc.K8sClient.AppsV1(), podsUnderTestLabelsObjects, data.Namespaces)
 	// Find ClusterRoleBindings
 	clusterRoleBindings, err := getClusterRoleBindings()
 	if err != nil {
