+++ pkg/autodiscover/autodiscover_test.go
@@ -17,6 +17,7 @@
 package autodiscover
 
 import (
+	"reflect"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -103,6 +104,39 @@ func TestBuildLabelKeyValue(t *testing.T) {
 		k, v := buildLabelKeyValue(tc.testLabel)
 		assert.Equal(t, tc.expectedKey, k)
 		assert.Equal(t, tc.expectedVal, v)
+	}
+}
+
+func Test_createLabels(t *testing.T) {
+	type args struct {
+		labelStrings []string
+	}
+	tests := []struct {
+		name             string
+		args             args
+		wantLabelObjects []labelObject
+	}{
+		{
+			name:             "ok",
+			args:             args{labelStrings: []string{"test-network-function.com/generic: target"}},
+			wantLabelObjects: []labelObject{{LabelKey: "test-network-function.com/generic", LabelValue: "target"}},
+		},
+		{
+			name:             "ok1",
+			args:             args{labelStrings: []string{"test-network-function.com/generic   : 1"}},
+			wantLabelObjects: []labelObject{{LabelKey: "test-network-function.com/generic", LabelValue: "1"}},
+		},
+		{
+			name: "nok",
+			args: args{labelStrings: []string{"test-network-function.com/generic= target"}},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if gotLabelObjects := createLabels(tt.args.labelStrings); !reflect.DeepEqual(gotLabelObjects, tt.wantLabelObjects) {
+				t.Errorf("createLabels() = %v, want %v", gotLabelObjects, tt.wantLabelObjects)
+			}
+		})
 	}
 }
