+++ pkg/configuration/configuration.go
@@ -105,11 +105,6 @@ type ContainerImageIdentifier struct {
 	Digest string `yaml:"digest" json:"digest"`
 }
 
-type LabelObject struct {
-	LabelKey   string
-	LabelValue string
-}
-
 // TestConfiguration provides test related configuration
 type TestConfiguration struct {
 	// targetNameSpaces to be used in
@@ -120,10 +115,6 @@ type TestConfiguration struct {
 	PodsUnderTestLabels []string `yaml:"podsUnderTestLabels,omitempty" json:"podsUnderTestLabels,omitempty"`
 	// labels identifying operators unde test
 	OperatorsUnderTestLabels []string `yaml:"operatorsUnderTestLabels,omitempty" json:"operatorsUnderTestLabels,omitempty"`
-	// Parsed labels identifying pods under test
-	PodsUnderTestLabelsObjects []LabelObject `yaml:"-" json:"-"`
-	// Parsed labels identifying operators under test
-	OperatorsUnderTestLabelsObjects []LabelObject `yaml:"-" json:"-"`
 	// CertifiedContainerInfo is the list of container images to be checked for certification status.
 	CertifiedContainerInfo []ContainerImageIdentifier `yaml:"certifiedcontainerinfo,omitempty" json:"certifiedcontainerinfo,omitempty"`
 	// CertifiedOperatorInfo is list of operator bundle names that are queried for certification status.
