+++ pkg/configuration/utils_test.go
@@ -15,41 +15,3 @@
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 package configuration
-
-import (
-	"reflect"
-	"testing"
-)
-
-func Test_createLabels(t *testing.T) {
-	type args struct {
-		labelStrings []string
-	}
-	tests := []struct {
-		name             string
-		args             args
-		wantLabelObjects []LabelObject
-	}{
-		{
-			name:             "ok",
-			args:             args{labelStrings: []string{"test-network-function.com/generic: target"}},
-			wantLabelObjects: []LabelObject{{LabelKey: "test-network-function.com/generic", LabelValue: "target"}},
-		},
-		{
-			name:             "ok1",
-			args:             args{labelStrings: []string{"test-network-function.com/generic   : 1"}},
-			wantLabelObjects: []LabelObject{{LabelKey: "test-network-function.com/generic", LabelValue: "1"}},
-		},
-		{
-			name: "nok",
-			args: args{labelStrings: []string{"test-network-function.com/generic= target"}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if gotLabelObjects := createLabels(tt.args.labelStrings); !reflect.DeepEqual(gotLabelObjects, tt.wantLabelObjects) {
-				t.Errorf("createLabels() = %v, want %v", gotLabelObjects, tt.wantLabelObjects)
-			}
-		})
-	}
-}
